Logistic Regression with the Lasso
Function for building logistic Regression from scratch.
Tune the model by cross-validation.
Use the Lasso as model selection methods
The File is conditioned for package generating, with description,
import, export, examples.
Test Cases is available at the beginning of the file @examples
Some Visualizations of the model are built in test cases, also see the
@examples